Extract DynamoDB persister
