+++ lib/payload-persister.js
@@ -1,78 +1,23 @@
 "use strict";
 
-var url = require("url"),
+var fs = require("fs"),
+    path = require("path"),
+    url = require("url"),
     util = require("util");
 
-var AWS = require("aws-sdk");
-
-AWS.config.update({
-  region: process.env.AWS_DEFAULT_REGION || AWS.config.region || "us-east-1"
-});
-
-var dynamodb = new AWS.DynamoDB();
-
-// TODO load modules
 // TODO prioritize
 // TODO add a function that determines whether a given persister is appropriate
-var persisters = {
-  "dynamodb:": {
-    // Assumptions:
-    // key attribute is named "key" and is a String
-    // "result" attribute exists, is a String, and contains JSON
-    load: function(uri, callback) {
-      var tableName = uri.hostname,
-          key = uri.pathname;
+var persisters = fs.readdirSync(path.join(__dirname, "persisters"))
+  .map(function(x) {
+    var persister = require(path.join(__dirname, "activities", x));
 
-      return dynamodb.getItem({
-        TableName: tableName,
-        Key: {
-          key: {
-            S: key
-          }
-        },
-        ConsistentRead: true
-      }, function(err, data) {
-        if (err) {
-          return callback(err);
-        }
+    return [persister.protocol, persister];
+  })
+  .reduce(function(obj, data) {
+    obj[data[0]] = data[1];
 
-        try {
-          return callback(null, JSON.parse(data.Item.result.S));
-        } catch (err) {
-          return callback(err);
-        }
-      });
-    },
-
-    save: function(key, payload, callback) {
-      // TODO make this configurable (use the SWF domain name?)
-      var tableName = "test";
-
-      return dynamodb.putItem({
-        TableName: tableName,
-        Item: {
-          key: {
-            S: key
-          },
-          result: {
-            S: payload
-          },
-          createdAt: {
-            S: new Date().toISOString()
-          }
-        }
-      }, function(err, data) {
-        if (err) {
-          return callback(err);
-        }
-
-        return callback(null, {
-          uri: util.format("dynamodb://%s/%s", tableName, taskToken)
-        });
-      });
-    }
-  }
-};
+    return obj;
+  }, {});
 
 module.exports = {
   load: function(payload, callback) {
